Feat: Criado excecoes de servico customizadas do delete

+++ src/main/java/com/fabiomorais/desc/services/EmployeeService.java
@@ -3,12 +3,15 @@ package com.fabiomorais.desc.services;
 import com.fabiomorais.desc.dto.EmployeeDTO;
 import com.fabiomorais.desc.entities.Employee;
 import com.fabiomorais.desc.repositories.EmployeeRepository;
+import com.fabiomorais.desc.services.exception.DatabaseException;
 import com.fabiomorais.desc.services.exception.ResourceNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -52,9 +55,17 @@ public class EmployeeService {
         }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(EmployeeDTO dto, Employee entity) {
